paymentapi/requirements.txt
coverage
Django
django-nose
djangorestframework
nose
requests
apscheduler
psycopg2-binary
django-environ
